# docs/swagger.yaml
basePath: /
definitions:
  models.Purchase:
    properties:
      articleID:
        type: integer
      lieferant:
        type: string
      menge:
        type: integer
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: justin.kromlinger@stud.htwk-leipzig.de
    name: Justin Kromlinger
    url: https://hashworks.net
  description: Solution for 'Einkauf' backend task of https://sites.google.com/relaxdays.de/hackathon-relaxdays/startseite#h.klg8hathdmsn
  license:
    name: GNU Affero General Public License v3
    url: https://gnu.org/licenses/agpl.html
  title: Purchase Backend Task
  version: "1.0"
paths:
  /purchase:
    post:
      consumes:
      - application/json
      parameters:
      - description: Purchase to save
        in: body
        name: purchase
        required: true
        schema:
          $ref: '#/definitions/models.Purchase'
      responses:
        "204":
          description: ""
        "400":
          description: Invalid purchase
          schema:
            type: ""
      summary: Save a purchase
      tags:
      - Purchase
  /purchases:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Purchase'
            type: array
      summary: Returns all saved purchases
      tags:
      - Purchase
  /purchasesForArticle:
    get:
      parameters:
      - description: ID of article to query
        in: query
        name: x
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Purchase'
            type: array
      summary: Returns all saved purchases for a given article
      tags:
      - Purchase
  /searchLieferant:
    get:
      parameters:
      - description: Vendor query
        in: query
        name: x
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Purchase'
            type: array
      summary: Returns all saved purchases with a similar vendor
      tags:
      - Purchase
swagger: "2.0"